+++ DB/DBZ/standard/business/src/main/java/jp/co/ndensan/reams/db/dbz/business/core/IryohokenKanyuJokyoIdentifier.java
@@ -7,7 +7,6 @@ package jp.co.ndensan.reams.db.dbz.business.core;
 
 import java.io.Serializable;
 import jp.co.ndensan.reams.uz.uza.biz.ShikibetsuCode;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import lombok.Value;
 
 /**
@@ -17,7 +16,7 @@ import lombok.Value;
 public class IryohokenKanyuJokyoIdentifier implements Serializable {
 
     private final ShikibetsuCode 識別コード;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -26,7 +25,7 @@ public class IryohokenKanyuJokyoIdentifier implements Serializable {
      * @param 履歴番号 履歴番号
      */
     public IryohokenKanyuJokyoIdentifier(ShikibetsuCode 識別コード,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.識別コード = 識別コード;
         this.履歴番号 = 履歴番号;
     }
